+++ Change_Log.md
@@ -24,7 +24,6 @@ Day 4 (12/3/2025):
 - Nodes are now navigated by clicking on the node name
 - Inventory can be navigated by pressing keys 1-8 (I want to change this so that it is clickable)
 
-
 Day 5 (12/4/2025):
 - Added more images and another image folder
 - Added Has_Item function
@@ -32,7 +31,17 @@ Day 5 (12/4/2025):
 - Created separate player class and moved logic over from main
 - Added health and damage logic
 
-
-
-
 Day 6 (12/5/2025):
+- Added event and interaction logic
+- Health bar added to screen
+- Gameover screen when you hit 0 health
+- Organized files into structured folders
+- Added background and font to main menu
+- Added music
+
+Day 7 (12/6/2025):
+- Styled the GUI and HUD 
+- Shelf wobble minigame added for node 1
+- Added can images to sprites folder for gold and normal cans
+- Broke up the explore file into smaller files
+-